backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain>=0.2.6
langchain-community>=0.2.6
langchain-chroma>=0.1.2
chromadb>=0.5.3
ollama>=0.2.1
pymupdf>=1.24.5
python-docx>=1.1.2
pandas>=2.2.2
openpyxl>=3.1.4
pillow>=10.3.0
python-multipart>=0.0.9
pydantic>=2.7.4
aiosqlite>=0.20.0
httpx>=0.27.0
python-dotenv>=1.0.1
numexpr>=2.9.0
